MB-54303 : Failure modes on service crash
This PR adds logic for some common failure cases:
-  Crash after download: restart a bucket state tracker
routine if crashed before the feeds could be added
after resume.
- Crash during download: restart a bucket state tracker
routine based on the premise that the bucket will be
deleted eventually.
- Crash during/after upload: restart a bucket state
tracker routine since the bucket will eventually be
changed back to pre-hibernation state.
- Pause fails midway: cancels upload context.
  Resets pindex paths to add feeds from within the
  bucket state tracker routine when bucket comes online.
- Resume fails during download: deletes indexes linked to
the bucket.

Also, adds rev. numbers for prepare tasks.

Reviewed-on: https://review.couchbase.org/c/cbgt/+/182180
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aditi Ahuja <aditi.ahuja@couchbase.com>

+++ pindex.go
@@ -39,10 +39,10 @@ type PIndex struct {
 	SourceUUID       string     `json:"sourceUUID"`
 	SourceParams     string     `json:"sourceParams"`
 	SourcePartitions string     `json:"sourcePartitions"`
+	HibernationPath  string     `json:"hibernationPath"`
 	Path             string     `json:"-"` // Transient, not persisted.
 	Impl             PIndexImpl `json:"-"` // Transient, not persisted.
 	Dest             Dest       `json:"-"` // Transient, not persisted.
-	HibernationPath  string     `json:"-"`
 
 	sourcePartitionsMap map[string]bool // Non-persisted memoization.
 
@@ -117,7 +117,6 @@ func (p *PIndex) Clone() *PIndex {
 			SourcePartitions:    p.SourcePartitions,
 			sourcePartitionsMap: p.sourcePartitionsMap,
 			Path:                p.Path,
-			HibernationPath:     p.HibernationPath,
 			Impl:                p.Impl,
 			Dest:                p.Dest,
 			closed:              p.closed,
@@ -144,6 +143,8 @@ func restartPIndex(mgr *Manager, pindex *PIndex) {
 	mgr.Kick("restart-pindex:" + pindex.Name)
 }
 
+var ErrTerminatedDownload = fmt.Errorf("pindex: case of abruptly terminated download")
+
 // Creates a pindex, including its backend implementation structures,
 // and its files.
 func NewPIndex(mgr *Manager, name, uuid,
@@ -165,7 +166,7 @@ func NewPIndex(mgr *Manager, name, uuid,
 
 	impl, dest, err := NewPIndexImplEx(indexType, string(pBytes), sourceParams,
 		path, mgr, restart)
-	if err != nil {
+	if err != nil && err != ErrTerminatedDownload {
 		os.RemoveAll(path)
 		return nil, fmt.Errorf("pindex: new indexType: %s, indexParams: %s,"+
 			" path: %s, err: %s", indexType, indexParams, path, err)
@@ -281,10 +282,6 @@ func PIndexPath(dataDir, pindexName string) string {
 func PIndexNameFromPath(path string) string {
 	path = filepath.Base(path)
 	return path[0 : len(path)-len(pindexPathSuffix)]
-}
-
-func (pindex *PIndex) HibernationInProgress() bool {
-	return strings.HasPrefix(pindex.HibernationPath, HIBERNATE_TASK)
 }
 
 // Retrieves a pindex name from a pindex path.
